+++ python/tvm/topi/__init__.py
@@ -49,6 +49,7 @@ from . import generic
 from . import nn
 from . import x86
 from . import cuda
+from . import gpu
 from . import arm_cpu
 from . import mali
 from . import bifrost
